Updated surf map
This is a visualizer for wave heights in a local area. The colors represent the category of the conditions per the surf forecast provided by magic seaweed. 
The ultimate goal is to plot this onto a physical map, but in the meantime the map is visualized in open street maps.